- In run-time file I/O:  Remove fprintf statements.  Replace with cryptic error codes as is done for other run-time errors. - Correct input to xREAD_STRING:  It is a pointer to a string, not the value of a string. - Remove newline characters in buffer returned by fgets.  Use to manage Pascal EOLN function.

+++ insn16/prun/psysio.h
@@ -71,7 +71,7 @@ typedef enum openMode_e openMode_t;
  * Public Function Prototypes
  ***************************************************************************/
 
-void     pexec_InitializeFile(void);
-uint16_t pexec_sysio(struct pexec_s *st, uint16_t subfunc);
+void pexec_InitializeFile(void);
+int  pexec_sysio(struct pexec_s *st, uint16_t subfunc);
 
 #endif /* __PFILE_H */
